Moving and cleaning directories
