+++ shared/src/commonMain/kotlin/com/kevinschildhorn/fotopresenter/ui/screens/directory/composables/navbar/DirectoryNavigationItem.kt
@@ -1,8 +1,8 @@
-package com.kevinschildhorn.fotopresenter.ui.screens.directory.composables
+package com.kevinschildhorn.fotopresenter.ui.screens.directory.composables.navbar
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
@@ -11,30 +11,24 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import com.kevinschildhorn.atomik.color.base.composeColor
-import com.kevinschildhorn.fotopresenter.ui.screens.common.composables.AtomikIcon
 import com.kevinschildhorn.fotopresenter.ui.screens.common.composables.AtomikText
 import com.kevinschildhorn.fotopresenter.ui.screens.directory.DirectoryAtoms
-import compose.icons.EvaIcons
-import compose.icons.evaicons.Fill
-import compose.icons.evaicons.fill.Home
 
 @Composable
-fun DirectoryNavigationHome(onClick: () -> Unit) {
+fun DirectoryNavigationItem(title: String, onClick: () -> Unit) {
     val molecule = DirectoryAtoms.navigationItem
     Button(
         onClick = onClick,
-        modifier = Modifier.height(44.dp).width(44.dp).clip(RoundedCornerShape(molecule.radius.dp)),
+        modifier = Modifier.height(44.dp).clip(RoundedCornerShape(molecule.radius.dp)),
         colors =
         ButtonDefaults.buttonColors(
             backgroundColor = molecule.color.composeColor,
             disabledBackgroundColor = molecule.disabledColor.composeColor,
         ),
     ) {
-        AtomikIcon(
-            EvaIcons.Fill.Home,
+        AtomikText(
+            title,
             atom = molecule.textAtom,
-            contentDescription = "Home",
-            modifier = Modifier.size(44.dp)
         )
     }
 }
